requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
MarkupSafe==3.0.2
psutil==6.1.0
py4j==0.10.9.7
pyspark==3.5.3
requests==2.32.3
urllib3==2.2.3
websockets==14.1
Werkzeug==3.1.3
